rearrange the project and fix the imports.
