+++ src/quartic/common/utils.py
